refactor(test): relocate Jest config files outside __tests__ directories
Move Jest configuration files (jest.config.js, jest.setup.js, setup.ts) from
__tests__ directories to prevent Jest from attempting to run them as test
suites.

Changes:
- Move src/lib/audit-filter/__tests__/jest.config.js to src/lib/audit-filter/
- Move src/lib/audit-filter/__tests__/jest.setup.js to src/lib/audit-filter/
- Move src/lib/audit-filter/__tests__/setup.ts to src/lib/audit-filter/
- Update file references and import paths in moved configuration files
- Add proper ESLint globals for Jest environment in configuration files
- Mark T038 as completed in TODO.md

Resolves Jest test discovery issues where configuration files were being
executed as test suites.
